requirements.txt
fastmcp
pydantic-ai-slim[openai]
requests
python-dotenv